refactor: extract url generation logic into a helper
